+++ viteclientts/src/utils/useStableTranslation.tsx
@@ -1,0 +1,13 @@
+import { useCallback } from "react";
+import { useTranslation } from "react-i18next";
+
+export const useStableTranslation = () => {
+    const { t } = useTranslation();
+
+    const strableTranslate = useCallback(
+        (key: string) => t(key),
+        [t]
+    );
+
+    return strableTranslate;
+}